Move test from a11y to playwright
